requirements.txt
Django==2.2.5
psycopg2-binary
django-cors-headers
djangorestframework
djangorestframework-jwt
gunicorn
